requirements.txt
flask
redis
celery
netmiko
python-dotenv
requests
